% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFMModel2.R
\name{calculaterfm}
\alias{calculaterfm}
\title{calculateRFM}
\usage{
calculaterfm(x, re, fr, mo)
}
\arguments{
\item{re}{- Weight of recency}

\item{fr}{- Weight of frequency}

\item{mo}{- Weight of monetary}

\item{data}{- A data.table containing the transaction record details for every customer.}
}
\value{
Returns a data.data containing the recency, frequency and monetary core as well as the weighted final score and the group membership.
}
\description{
Description
Calclate the weighted RMF score: recency, frequency, and monetary.
}
\details{
\code{data} contains the transactional data. The dataset must contain a column labeled "Customer" that allows unique customer identification and a column labeled "TransDate", indicating the purchase date. The column "PurchAmount" specifies the total spending per purchase.
Return Value
}
