% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFMScores.R
\name{calculateRFMScores}
\alias{calculateRFMScores}
\title{calculateRFMScores}
\usage{
calculateRFMScores(data, weight_recency = 1, weight_frequency = 1,
  weight_monetary = 1)
}
\arguments{
\item{data}{- A data.table containing the transaction details for every customer. Transaction date has to be of the lubridate form "yyyy-mm-dd" and have name "TransDate".}

\item{r}{- Weight of recency.}

\item{f}{- Weight of frequency.}

\item{m}{- Weight of monetary value.}
}
\value{
Returns a data.table containing the recency, frequency and monetary
score as well as the weighted final score and the group membership.
}
\description{
Calculate a weighted RFM score: recency, frequency, and monetary value.
}
\details{
\code{data} contains the transactional data. The dataset must contain a
column labeled "Customer" that allows unique customer identification
and a column labeled "TransDate", indicating the purchase date. The column
"PurchAmount" specifies the total spending per purchase.
}
