% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFMfunction.R
\name{RFMfunction}
\alias{RFMfunction}
\title{Calculate weighted RFM scores based on recency, frequency and monetary values of customer purchases and sort customers to groups}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{data}{- database of customer purchases with details}

\item{weight_recency}{- weight of recency factor}

\item{weight_frequency}{- weight of frequency factor}

\item{weight_monetary}{- weight of monetary values of purchases}
}
\value{
Returns a data file with weighted scores of customers and their assigned group
}
\description{
Arguments
}
