% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFM_Function.R
\name{RFMfunction}
\alias{RFMfunction}
\title{RFMfunction}
\usage{
RFMfunction(d, weight_recency = 1, weight_frequency = 1, weight_monetary = 1)
}
\arguments{
\item{d}{A data table containing columns TransDate, Customer & PurchAmount}

\item{weight_recency}{the relative weight of the Recency variable in the overall score}

\item{weight_frequency}{the relative weight of the Frequency variable in the overall score}

\item{weight_monetary}{the relative weight of the Monetary variable in the overall score

Sum of weights must add to 100.

Returned values}
}
\value{
A data table containing the overall score of the RFM analysis plus the individual values for  \code{weight_recency}, \code{weight_frequency}
and \code{weight_monetary}
}
\description{
Description
This function creates an RFM table for customer analysis
}
\details{
Arguments
}
\examples{
RFMfunction(transactions, 60, 20, 20)
RFMfunction(customer_data, 30, 30, 40)

}
