% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFM.R
\name{RFMfunction}
\alias{RFMfunction}
\title{RFM function}
\usage{
RFMfunction(d, weight_recency = 1, weight_frequency = 1, weight_monetary = 1)
}
\arguments{
\item{arg1}{data set}
}
\description{
This function calculatethe RFM using weights and a data set as paremeters
}
\details{
\code {data} Data set must have a "TransDate" column, "Customer" column and a "PurchAmount" column
}
