% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFMfunction.R
\name{RFMfunction}
\alias{RFMfunction}
\title{calculating rfm
calculate recency, frequency, monetary scores for each customer}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{data}{Ensure that the weights add up to one
RFM measures
RFM scores
Overall RFM score
RFM group
Return final table}
}
\description{
calculating rfm
calculate recency, frequency, monetary scores for each customer
}
